Use phoenix sanitization when available
Phoenix uses a custom EEx engine to ensure that all values injected into
views are "safe" from XSS attacks. Since ratchet builds html attributes
dynamically, these values must be marked as safe for the correct view to
be rendered by Phoenix. However, it is desirable to keep the ratchet
library decoupled from Phoenix as it may be used in other contexts.

To be frank, this solution feels bad. Just because the Phoenix.HTML
module is loaded does not mean that it is being used in rendering the
template. Therefore, there is a chance that views rendered outside
Phoenix (but in its presence) will confusing be marked as "Phoenix
safe". For now, that's a risk we'll have to take... Would love to find a
better solution to this problem.