Mark generated attributes as safe
In order for ratchet to work with Phoenix, generated attributes must be
flagged as "raw". Otherwise they would be sanitized during rendering.
This has the unfortunately effect of coupling ratchet to phoenix... I
think we'll just be OK with this for now given that phoenix is probably
its only use case. Eventually we'll need to provide some sort of hook
that performs the special behavior in the presence of phoenix.