Add property support for non-map data
In the updated example the "bar" property is fetched from foo's data
(which is {nil, []}). Since foo's data is not a map, this would result
in an error. In this case, no data is provided for bar, so nil is
returned for that property. Further, we must update prepare/1 to handle
nil data. The default behavior of List.wrap is to return an empty list,
but in our case we want to actually return a list of just nil. That is
so that the default content is rendered once with nil data.