Remove Ratchet.Data module
After experimenting with this a bit, I don't think it's a useful
abstraction. Instead of making things easier, it just hides the fact
that data is a simple Map data structure. Further, there are performance
concerns with repeated updating maps and such when you could just be
coding one directly. In the end, maybe there is a more elegant
data-building strategy, but I don't think this is it.